33rd commit

+++ admin/repairs.php
@@ -102,17 +102,19 @@
                                                 Actions
                                             </button>
                                             <ul class="dropdown-menu" aria-labelledby="actionDropdown<?= $repair['id']; ?>">
-                                                <li>
-                                                    <a class="dropdown-item" href="repairs-edit.php?id=<?= $repair['id']; ?>">
-                                                        <i class="fas fa-edit"></i> Edit
-                                                    </a>
-                                                </li>
-                                                <li>
-                                                    <a class="dropdown-item" href="repairs-delete.php?id=<?= $repair['id']; ?>" 
-                                                    onclick="return confirm('Are you sure you want to delete this repair?');">
-                                                        <i class="fas fa-trash-alt"></i> Delete
-                                                    </a>
-                                                </li>
+                                                <?php if ($_SESSION['role'] == 'admin'): ?>
+                                                    <li>
+                                                        <a class="dropdown-item" href="repairs-edit.php?id=<?= $repair['id']; ?>">
+                                                            <i class="fas fa-edit"></i> Edit
+                                                        </a>
+                                                    </li>
+                                                    <li>
+                                                        <a class="dropdown-item" href="repairs-delete.php?id=<?= $repair['id']; ?>" 
+                                                        onclick="return confirm('Are you sure you want to delete this repair?');">
+                                                            <i class="fas fa-trash-alt"></i> Delete
+                                                        </a>
+                                                    </li>
+                                                <?php endif; ?>
                                                 <li>
                                                     <a class="dropdown-item" href="repairs-view.php?id=<?= $repair['id']; ?>">
                                                         <i class="fas fa-eye"></i> View
@@ -127,7 +129,7 @@
                     </table>
                 </div>
             <?php } else { ?>
-                <h4>No Repair Items Found
+                <h4>No Repair Items Found</h4>
             <?php } ?>
         </div>
     </div>
